[genesis] Refactor genesis code
