+++ chain/network/src/network_protocol/mod.rs
@@ -7,6 +7,7 @@ mod peer;
 mod proto_conv;
 mod state_sync;
 pub use edge::*;
+use near_primitives::genesis::GenesisId;
 use near_primitives::stateless_validation::chunk_endorsement::ChunkEndorsement;
 use near_primitives::stateless_validation::contract_distribution::ChunkContractAccesses;
 use near_primitives::stateless_validation::contract_distribution::ContractCodeRequest;
@@ -40,7 +41,7 @@ use near_async::time;
 use near_crypto::PublicKey;
 use near_crypto::Signature;
 use near_o11y::OpenTelemetrySpanExt;
-use near_primitives::block::{Approval, Block, BlockHeader, GenesisId};
+use near_primitives::block::{Approval, Block, BlockHeader};
 use near_primitives::challenge::Challenge;
 use near_primitives::epoch_sync::CompressedEpochSyncProof;
 use near_primitives::hash::CryptoHash;
