+++ chain/network/src/network_protocol/peer.rs
@@ -3,7 +3,7 @@
 /// All changes to this file should be reviewed.
 ///
 /// TODO: - document all types in this file
-use near_primitives::block::GenesisId;
+use near_primitives::genesis::GenesisId;
 use near_primitives::network::PeerId;
 use near_primitives::types::{AccountId, BlockHeight, ShardId};
 use near_schema_checker_lib::ProtocolSchema;
