+++ chain/network/src/network_protocol/testonly.rs
@@ -8,8 +8,9 @@ use crate::tcp;
 use crate::types::{AccountKeys, ChainInfo, Handshake, RoutingTableUpdate};
 use near_async::time;
 use near_crypto::{InMemorySigner, KeyType, SecretKey, Signer};
-use near_primitives::block::{Block, BlockHeader, GenesisId, genesis_chunks};
+use near_primitives::block::{Block, BlockHeader};
 use near_primitives::challenge::{BlockDoubleSign, Challenge, ChallengeBody};
+use near_primitives::genesis::genesis_chunks;
 use near_primitives::hash::CryptoHash;
 use near_primitives::network::{AnnounceAccount, PeerId};
 use near_primitives::num_rational::Ratio;
