+++ core/primitives/benches/serialization.rs
@@ -6,7 +6,8 @@ use bencher::{Bencher, black_box};
 use borsh::BorshDeserialize;
 use near_crypto::{KeyType, PublicKey, Signature};
 use near_primitives::account::Account;
-use near_primitives::block::{Block, genesis_chunks};
+use near_primitives::block::Block;
+use near_primitives::genesis::genesis_chunks;
 use near_primitives::hash::CryptoHash;
 use near_primitives::merkle::combine_hash;
 use near_primitives::test_utils::account_new;
