+++ core/primitives/src/genesis/block.rs
@@ -1,0 +1,58 @@
+use near_primitives_core::hash::CryptoHash;
+use near_primitives_core::types::{Balance, BlockHeight, ProtocolVersion};
+use near_time::Utc;
+
+use crate::block::{Block, BlockHeader};
+use crate::block_body::BlockBody;
+use crate::sharding::ShardChunkHeader;
+
+impl Block {
+    /// Returns genesis block for given genesis date and state root.
+    pub fn genesis(
+        genesis_protocol_version: ProtocolVersion,
+        chunks: Vec<ShardChunkHeader>,
+        timestamp: Utc,
+        height: BlockHeight,
+        initial_gas_price: Balance,
+        initial_total_supply: Balance,
+        next_bp_hash: CryptoHash,
+    ) -> Self {
+        let challenges = vec![];
+        let chunk_endorsements = vec![];
+        for chunk in &chunks {
+            assert_eq!(chunk.height_included(), height);
+        }
+        let vrf_value = near_crypto::vrf::Value([0; 32]);
+        let vrf_proof = near_crypto::vrf::Proof([0; 64]);
+        let body = BlockBody::new(
+            genesis_protocol_version,
+            chunks,
+            challenges,
+            vrf_value,
+            vrf_proof,
+            chunk_endorsements,
+        );
+        let header = BlockHeader::genesis(
+            genesis_protocol_version,
+            height,
+            Block::compute_state_root(body.chunks()),
+            body.compute_hash(),
+            Block::compute_chunk_prev_outgoing_receipts_root(body.chunks()),
+            Block::compute_chunk_headers_root(body.chunks()).0,
+            Block::compute_chunk_tx_root(body.chunks()),
+            body.chunks().len() as u64,
+            Block::compute_challenges_root(body.challenges()),
+            timestamp,
+            initial_gas_price,
+            initial_total_supply,
+            next_bp_hash,
+        );
+
+        Self::block_from_protocol_version(
+            genesis_protocol_version,
+            genesis_protocol_version,
+            header,
+            body,
+        )
+    }
+}
