+++ core/primitives/src/version.rs
@@ -17,6 +17,7 @@ use crate::upgrade_schedule::ProtocolUpgradeVotingSchedule;
 pub use near_primitives_core::types::ProtocolVersion;
 pub use near_primitives_core::version::MIN_SUPPORTED_PROTOCOL_VERSION;
 pub use near_primitives_core::version::PEER_MIN_ALLOWED_PROTOCOL_VERSION;
+pub use near_primitives_core::version::PROD_GENESIS_PROTOCOL_VERSION;
 pub use near_primitives_core::version::PROTOCOL_VERSION;
 pub use near_primitives_core::version::ProtocolFeature;
