+++ integration-tests/src/tests/network/stress_network.rs
@@ -11,7 +11,7 @@ use near_actix_test_utils::run_actix;
 use near_async::time;
 use near_network::tcp;
 use near_o11y::testonly::init_test_logger_allow_panic;
-use near_primitives::block::GenesisId;
+use near_primitives::genesis::GenesisId;
 
 use near_network::PeerManagerActor;
 use near_network::config;
