+++ test-loop-tests/src/setup/state.rs
@@ -14,8 +14,8 @@ use near_network::shards_manager::ShardsManagerRequestFromNetwork;
 use near_network::state_witness::PartialWitnessSenderForNetwork;
 use near_network::types::{HighestHeightPeerInfo, PeerInfo};
 use near_parameters::RuntimeConfigStore;
-use near_primitives::block::GenesisId;
 use near_primitives::epoch_manager::EpochConfigStore;
+use near_primitives::genesis::GenesisId;
 use near_primitives::network::PeerId;
 use near_primitives::types::AccountId;
 use near_primitives::upgrade_schedule::ProtocolUpgradeVotingSchedule;
